Played around with worker
still not complete

+++ Examples/majordomo/worker/src/reduction.rs
@@ -1,15 +1,15 @@
-pub fn calc_avg(thread_num: usize, data: Vec<u64>) {
+pub fn calc_avg(thread_num: usize, data: Vec<i32>) {
     let mut sum = 0;
     for value in data.iter()
     {
         sum = sum + value;
     }
-    let avg = sum / data.len() as u8;
+    let avg = sum / data.len() as i32;
 
     println!("Thread: {:?} -> Average: {:?}", thread_num, avg);
 }
 
-pub fn calc_sum(thread_num: usize, data: Vec<u64>) {
+pub fn calc_sum(thread_num: usize, data: Vec<i32>) {
     let mut sum = 0;
     for value in data.iter()
     {
@@ -19,7 +19,7 @@ pub fn calc_sum(thread_num: usize, data: Vec<u64>) {
     println!("Thread: {:?} -> Sum: {:?}", thread_num, sum);
 }
 
-pub fn calc_min(thread_num: usize, data: Vec<u64>) {
+pub fn calc_min(thread_num: usize, data: Vec<i32>) {
     let mut min = data[0];
     for value in data.iter()
     {
@@ -32,7 +32,7 @@ pub fn calc_min(thread_num: usize, data: Vec<u64>) {
     println!("Thread: {:?} -> Min: {:?}", thread_num, min);
 }
 
-pub fn calc_max(thread_num: usize, data: Vec<u64>) {
+pub fn calc_max(thread_num: usize, data: Vec<i32>) {
     let mut max = data[0];
     for value in data.iter()
     {
